task2 : further to listing all pets Added a new pet controller.
