+++ src/test/java/org/springframework/samples/petclinic/service/ClinicServiceTests.java
@@ -36,6 +36,7 @@ import org.springframework.samples.petclinic.visit.Visit;
 import org.springframework.samples.petclinic.visit.VisitRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.samples.petclinic.service.EntityUtils;
 
 /**
  * Integration test of the Service and the Repository layer.
